Fixed renaming mismatch if files in db folder weren't the same as the db (i.e. Financial_Questions in TAP_Portfolio).
git-svn-id: svn://107.170.10.188:8080/Semoss/trunk@4622 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 493cb726fb4894cd8a23f1b114f5236d384e86ad

+++ src/prerna/ui/comparison/specific/tap/DBRenamer.java
@@ -149,11 +149,12 @@ public class DBRenamer
 			public void actionPerformed(ActionEvent e)
 			{
 				String folderDirectory = importFolderNameField.getText();
+				String dbName = folderDirectory.substring(folderDirectory.indexOf("db\\") + 3);
 				File folder = new File(folderDirectory);
 				File[] listOfFiles = folder.listFiles();
 				for (int i = 0; i < listOfFiles.length; i++)
 				{
-					if (listOfFiles[i].isFile())
+					if (listOfFiles[i].isFile() && listOfFiles[i].getName().contains(dbName))
 					{
 						File f = new File(folderDirectory + "\\" + listOfFiles[i].getName());
 						f.renameTo(new File(folderDirectory + "\\old" + listOfFiles[i].getName()));
